app/requirements.txt
fastapi>=0.100.0
uvicorn
python-multipart
PyMuPDF
pydantic>=2.0
langchain>=0.1.0
python-dotenv
langchain-community